created a minted crowdsale behaviour to be tested with both mintable and rbacmintable token
