+++ test/crowdsale/MintedCrowdsale.test.js
@@ -1,61 +1,45 @@
+import shouldBehaveLikeMintedCrowdsale from './MintedCrowdsale.behaviour';
 import ether from '../helpers/ether';
 
 const BigNumber = web3.BigNumber;
 
-const should = require('chai')
-  .use(require('chai-as-promised'))
-  .use(require('chai-bignumber')(BigNumber))
-  .should();
-
 const MintedCrowdsale = artifacts.require('MintedCrowdsaleImpl');
 const MintableToken = artifacts.require('MintableToken');
+const RBACMintableToken = artifacts.require('RBACMintableToken');
 
 contract('MintedCrowdsale', function ([_, investor, wallet, purchaser]) {
   const rate = new BigNumber(1000);
-  const value = ether(42);
+  const value = ether(5);
 
-  const expectedTokenAmount = rate.mul(value);
+  describe('using MintableToken', function () {
+    beforeEach(async function () {
+      this.token = await MintableToken.new();
+      this.crowdsale = await MintedCrowdsale.new(rate, wallet, this.token.address);
+      await this.token.transferOwnership(this.crowdsale.address);
+    });
 
-  beforeEach(async function () {
-    this.token = await MintableToken.new();
-    this.crowdsale = await MintedCrowdsale.new(rate, wallet, this.token.address);
-    await this.token.transferOwnership(this.crowdsale.address);
-  });
-
-  describe('accepting payments', function () {
     it('should be token owner', async function () {
       const owner = await this.token.owner();
       owner.should.equal(this.crowdsale.address);
     });
 
-    it('should accept payments', async function () {
-      await this.crowdsale.send(value).should.be.fulfilled;
-      await this.crowdsale.buyTokens(investor, { value: value, from: purchaser }).should.be.fulfilled;
-    });
+    shouldBehaveLikeMintedCrowdsale([_, investor, wallet, purchaser], rate, value);
   });
 
-  describe('high-level purchase', function () {
-    it('should log purchase', async function () {
-      const { logs } = await this.crowdsale.sendTransaction({ value: value, from: investor });
-      const event = logs.find(e => e.event === 'TokenPurchase');
-      should.exist(event);
-      event.args.purchaser.should.equal(investor);
-      event.args.beneficiary.should.equal(investor);
-      event.args.value.should.be.bignumber.equal(value);
-      event.args.amount.should.be.bignumber.equal(expectedTokenAmount);
+  describe('using RBACMintableToken', function () {
+    const ROLE_MINTER = 'minter';
+
+    beforeEach(async function () {
+      this.token = await RBACMintableToken.new();
+      this.crowdsale = await MintedCrowdsale.new(rate, wallet, this.token.address);
+      await this.token.addMinter(this.crowdsale.address);
     });
 
-    it('should assign tokens to sender', async function () {
-      await this.crowdsale.sendTransaction({ value: value, from: investor });
-      let balance = await this.token.balanceOf(investor);
-      balance.should.be.bignumber.equal(expectedTokenAmount);
+    it('should have minter role on token', async function () {
+      const isMinter = await this.token.hasRole(this.crowdsale.address, ROLE_MINTER);
+      isMinter.should.equal(true);
     });
 
-    it('should forward funds to wallet', async function () {
-      const pre = web3.eth.getBalance(wallet);
-      await this.crowdsale.sendTransaction({ value, from: investor });
-      const post = web3.eth.getBalance(wallet);
-      post.minus(pre).should.be.bignumber.equal(value);
-    });
+    shouldBehaveLikeMintedCrowdsale([_, investor, wallet, purchaser], rate, value);
   });
 });
